+++ package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "class-variance-authority": "^0.7.0",
+    "lodash": "^4.17.21",
     "next": "14.2.4",
     "react": "^18",
     "react-dom": "^18",
